refactor(gtw): unify logging to single log.js in ~/.gtw/logs/
- Create utils/log.js with unified log() and logParseFailure()
- index.js now uses log() from utils/log.js (was inline dbg())
- Delete old utils/debug-log.js (merged into log.js)
- All logs now go to ~/.gtw/logs/gtw.log
